+++ Sora/PackageInfo.swift
@@ -9,16 +9,16 @@ public enum SDKInfo {
  */
 public enum WebRTCInfo {
     /// WebRTC フレームワークのバージョン
-    public static let version = "M104"
+    public static let version = "M105"
 
     /// WebRTC フレームワークのコミットポジション
-    public static let commitPosition = "8"
+    public static let commitPosition = "0"
 
     /// WebRTC フレームワークのメンテナンスバージョン
     public static let maintenanceVersion = "0"
 
     /// WebRTC フレームワークのソースコードのリビジョン
-    public static let revision = "06aea31d10f860ae4236e3422252557762d39188"
+    public static let revision = "dc5cf31cad576376abd3aa6306169453cfd85ba5"
 
     /// WebRTC フレームワークのソースコードのリビジョン (短縮版)
     public static var shortRevision: String {